[nvim] symlink the whole directory
Previously, we symlinked individual files, so as to avoid symlinking
init.vim, which we no longer use. By moving it to `old/`, we can fix
this issue. We also (I believe) no longer use a python provider for any
plugins, so we can move the pipenv stuff to old as well.