Properly setup cmp_git
The github filetype will make the plugin active for neovim buffers
opened from the github command line, because I overrode the default
editor for `gh`. In particular, there's a PR I wrote that I'm waiting to
be merged in cmp-git that implements support for dot-separated
filetypes, so we can have markdown support and also treat it as a github
file.